Added inactive water type (HRU_TYPE=4)
Cells that flow to inactive water without going to a stream will be
grouped into a single subbasin.  All other cells that flow out of the
model without going through a stream will be kept but assigned to a
separate subbasin.  The HRU_TYPE=4 is only used in the HRU_TYPE_IN field
and is not set in the final HRU_TYPE field.  The general approach was to
create subbasin points at all cells that are exiting the model study
area.