Mods so that cells at edge of study area are handled correctly
Was having problems with active edge cells flowing out of the study area
(because it was steeper out) even though majority of cell was in.  Added
a flag so so that ArcGIS only calculates flow
direction/accumulation/fills sinks for active cells (in the study area).
This is more similar to was CRT will do also.  Currently this needs to
be turned on (mask_inactive_cells_flag).  This is applied independently
of the set_inactive_water flag.  I still need to apply this for 4-way
sink filling (since I am not using the ArcGIS fill function).