Update OUTFLOW parameter
The OUTFLOWHRU (formerly called OUTFLOW) field in hru_params.shp is now
set to 1 for all cells that exit the model domain or exit to an inactive
field.  This parameter is also now independent of the gauges.  This
parameter is used to set the cells in the CRT OUTFLOW_HRU.DAT file.